pyproject.toml
[tool.poetry]
name = "acels"
version = "0.1.0"
description = "Levitation Feedback Control Loop Software"
authors = ["Zalan Nemeth <zalan.nemeth@riverlane.com>"]

[tool.poetry.dependencies]
black = "^22.3"
python = ">=3.9,<3.12"
pytest = {version = "^7.2.0", optional = true}
pytest-cov = {version = "^4.0.0", optional = true}
pytest-random-order = {version = "^1.0", optional = true}
pylint = {version = "^2.15.9", optional = true}
pycodestyle = {version = "^2.8.0", optional = true}
mypy = {version = "^1.0.0", optional = true}
autopep8 = {version = "^2.0.1", optional = true}
pip-licenses = {version = "^3.5.3", optional = true}
pre-commit = {version = "^2.20.0", optional = true}
numpy = ">=1.22.4,<2"
tensorflow = "2.15"
tensorflow-io-gcs-filesystem = { version = "0.34.0"}
matplotlib = "^3.7.2"
pandas = "^2.1.3"
scikit-learn = "^1.4.1.post1"
pyserial = "^3.5"

[tool.poetry.extras]
test = ["pytest", "pytest-cov", "pytest-random-order"]
lint = ["pylint", "mypy", "pycodestyle", "autopep8", "pre-commit"]
licences = ["pip-licenses"]

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
